GVT-2708: Separate duplicate track entry into its own component. Also unify translation for "marked"
