+++ infra/src/main/resources/i18n/translations.fi.json
@@ -642,9 +642,9 @@
             "location-track": "Raide \"{{trackName}}\"",
             "errors": "Virheet ({{count}}):",
             "infos": "Huomiot ({{count}}):",
-            "implicit-duplicate-tooltip": "Raidetta \"{{trackName}}\" ei ole merkattu raiteen \"{{otherTrackName}}\" duplikaatiksi, mutta raiteilla on päällekkäistä geometriaa",
-            "non-overlapping-duplicate-tooltip": "Raide \"{{trackName}}\" on merkattu raiteen \"{{otherTrackName}}\" duplikaatiksi, mutta raiteilla ei ole päällekkäistä geometriaa",
-            "overlapping-duplicate-of-different-track-tooltip": "Raide \"{{implicitDuplicateName}}\" on merkattu eri raiteen duplikaatiksi (\"{{explicitDuplicateName}}\",) mutta sillä on päällekkäistä geometriaa myös raiteen \"{{trackName}}\" kanssa",
+            "implicit-duplicate-tooltip": "Raidetta \"{{trackName}}\" ei ole merkitty raiteen \"{{otherTrackName}}\" duplikaatiksi, mutta raiteilla on päällekkäistä geometriaa",
+            "non-overlapping-duplicate-tooltip": "Raide \"{{trackName}}\" on merkitty raiteen \"{{otherTrackName}}\" duplikaatiksi, mutta raiteilla ei ole päällekkäistä geometriaa",
+            "overlapping-duplicate-of-different-track-tooltip": "Raide \"{{implicitDuplicateName}}\" on merkitty eri raiteen duplikaatiksi (\"{{explicitDuplicateName}}\",) mutta sillä on päällekkäistä geometriaa myös raiteen \"{{trackName}}\" kanssa",
             "has-duplicates": "Duplikaattiraiteet",
             "not-a-duplicate": "Ei duplikaatti",
             "change-info-heading": "Lokitiedot",
